adição de verificação de application-key; criada função de verificação

+++ app/Http/Controllers/api/DogController.php
@@ -4,10 +4,26 @@ namespace App\Http\Controllers\api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Dog;
+use Exception;
 use Illuminate\Http\Request;
+use PhpParser\Node\Stmt\Return_;
+
+function verificacao($request){
+        if($request->header('application-key') == null){
+            return ['error' => ['id' => 1001,'description' => 'Api key não recebida!']];
+        }else{
+            if($request->header('application-key') !== "minhakey123"){
+                return ['error' => ['id' => 1002,'description' => 'Key informada invalida']];
+            }else{
+                return true;
+            }
+        }
+    }
 
 class DogController extends Controller
 {
+
+    
     /**
      * Display a listing of the resource.
      *
@@ -26,20 +42,30 @@ class DogController extends Controller
      */
     public function store(Request $request)
     {
-        if($request->header('application-key') == null){
-            $result = ['error' => ['id' => 1001,'description' => 'Api key não recebida!']];
-        }else{
-            if($request->header('application-key') !== "minhakey123"){
-                $result = ['error' => ['id' => 1002,'description' => 'Key informada invalida']];
-            }else{
-                if($request->raca !== null && $request->nome !== null){
+        $dadosverificacao = verificacao($request);
+
+        if($dadosverificacao == "true"){
+            if($request->raca !== null && $request->nome !== null){
+                try{
                     $dog = Dog::create($request->all());
                     $result = $dog;
-                }else{
-                    $result = ['error' => ['id' => 1000,'description' => 'Necessario Informar nome e raca!']];
+                }catch(Exception $e){
+                    if($e->getCode() == "23000"){
+                        $result = ['error' => ['id' => $e->getCode(),'description' => 'Nome informado já existe']];
+                    }else{
+                        $errrortrat = explode(":", $e->getPrevious()->getMessage());
+                        $result = ['error' => ['id' => $e->getCode(),'description' => $errrortrat[1]]];
+                    }
                 }
+                
+            }else{
+                $result = ['error' => ['id' => 1000,'description' => 'Necessario Informar nome e raca!']];
             }
-        }  
+        }else{
+            $result = $dadosverificacao;
+        }
+          
+        
 
         //return $request->header();
         return json_encode($result);
@@ -65,8 +91,36 @@ class DogController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $dog = Dog::findOrFail($id);
-        $dog->update($request->all());
+        $dadosverificacao = verificacao($request);
+
+        if($dadosverificacao == "true"){
+            if($request->raca !== null && $request->nome !== null){
+                try{
+                    $dog = Dog::findOrFail($id);
+                    $dog->update($request->all());
+                    $result = $dog;
+                }catch(Exception $e){
+                    if($e->getCode() == "23000"){
+                        $result = ['error' => ['id' => $e->getCode(),'description' => 'Nome informado já existe']];
+                    }else{
+                        $errrortrat = explode(":", $e->getPrevious()->getMessage());
+                        $result = ['error' => ['id' => $e->getCode(),'description' => $errrortrat[1]]];
+                    }
+                }
+                
+            }else{
+                $result = ['error' => ['id' => 1000,'description' => 'Necessario Informar nome e raca!']];
+            }
+        }else{
+            $result = $dadosverificacao;
+        }
+          
+        
+        return $result;
+        //return $request->header();
+        //return json_encode($result);
+        
+        
     }
 
     /**
@@ -75,9 +129,32 @@ class DogController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $dog = Dog::findOrFail($id);
-        $dog->delete();
+        $dadosverificacao = verificacao($request);
+
+        if($dadosverificacao == "true"){
+            try{
+                $dog = Dog::findOrFail($id);
+                $dog->delete();
+                $result = $dog;
+            }catch(Exception $e){
+                if($e->getCode() == "23000"){
+                    $result = ['error' => ['id' => $e->getCode(),'description' => 'Nome informado já existe']];
+                }else{
+                    $errrortrat = explode(":", $e->getPrevious()->getMessage());
+                    $result = ['error' => ['id' => $e->getCode(),'description' => $errrortrat[1]]];
+                }
+            }
+        }else{
+            $result = $dadosverificacao;
+        }
+          
+        
+
+        //return $request->header();
+        return json_encode($result);
+
+        
     }
 }
